mealprep/requirements.txt
asgiref==3.6.0
Django==4.1.5
django-recaptcha==3.0.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
gunicorn==20.1.0
requests==2.28.2

